+++ L02/Acme-Madruga/src/main/java/converters/FloatToStringConverter.java
@@ -4,13 +4,13 @@ package converters;
 import org.springframework.core.convert.converter.Converter;
 import org.springframework.stereotype.Component;
 
-import domain.BigDecimal;
+import domain.Float;
 
 @Component
-public class FloatToStringConverter implements Converter<domain.BigDecimal, String> {
+public class FloatToStringConverter implements Converter<domain.Float, String> {
 
 	@Override
-	public String convert(final domain.BigDecimal floatt) {
+	public String convert(final domain.Float floatt) {
 		String result;
 
 		if (floatt == null)
